requirements.txt
-e .
aws-cdk.aws_ec2
aws-cdk.aws_elasticloadbalancingv2